# save distfile and use it if it exists