Now cpm save dists in ~/perl-cpm/cache:
```
> find ~/.perl-cpm/cache -type f
/Users/skaji/.perl-cpm/cache/authors/id/A/AD/ADAMK/Class-Inspector-1.28.tar.gz
/Users/skaji/.perl-cpm/cache/authors/id/A/AD/ADAMK/Params-Util-1.07.tar.gz
/Users/skaji/.perl-cpm/cache/authors/id/A/AR/ARISTOTLE/Hash-MultiValue-0.16.tar.gz
/Users/skaji/.perl-cpm/cache/authors/id/C/CJ/CJM/IO-HTML-1.001.tar.gz
/Users/skaji/.perl-cpm/cache/authors/id/C/CO/CORION/parent-0.234.tar.gz
/Users/skaji/.perl-cpm/cache/authors/id/D/DA/DAGOLDEN/File-pushd-1.009.tar.gz
...
```
And the cache exists, cpm use it instead of fetching it from CPAN.
This makes cpm even faster!

Without cache
```
> perl -Ilib script/cpm install Plack
perl -Ilib script/cpm install Plack  18.85s user 5.00s system 210% cpu 11.334 total
```

With cache
```
> perl -Ilib script/cpm install Plack
perl -Ilib script/cpm install Plack  19.39s user 5.07s system 290% cpu 8.430 total
```

11.334sec -> 8.430sec :congratulations: